requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
drf-yasg==1.21.5